+++ app/docs/layout.tsx
@@ -1,4 +1,5 @@
 import { Leftbar } from "@/components/leftbar";
+import { MDXProviderWrapper } from "@/components/markdown/mdx-provider";
 
 export default function DocsLayout({
   children,
@@ -8,7 +9,11 @@ export default function DocsLayout({
   return (
     <div className="flex items-start gap-8">
       <Leftbar key="leftbar" />
-      <div className="flex-[5.25]">{children}</div>
+      <div className="flex-[5.25]">
+        <MDXProviderWrapper>
+          {children}
+        </MDXProviderWrapper>
+      </div>
     </div>
   );
 }